+++ demo/pom.xml
@@ -22,7 +22,19 @@
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
 
-
+	<dependency>
+    <groupId>io.jsonwebtoken</groupId>
+    <artifactId>jjwt</artifactId>
+    <version>0.6.0</version>
+	</dependency>
+	
+	
+	<dependency>
+    <groupId>io.jsonwebtoken</groupId>
+    <artifactId>jjwt-api</artifactId>
+    <version>0.10.5</version>
+	</dependency>
+	
 
 
 	<dependency>
@@ -82,6 +94,7 @@
 			<artifactId>reactor-test</artifactId>
 			<scope>test</scope>
 		</dependency>
+
 		
 	</dependencies>
