--- dev/demo/Dockerfile
@@ -1,4 +1,0 @@
-FROM openjdk:15
-COPY target/demo-0.0.1-SNAPSHOT.jar app.jar
-EXPOSE 8080
-ENTRYPOINT ["java","-jar","app.jar"]
