--- dev/demo/docker-compose.yml
@@ -1,24 +1,0 @@
-version: '2'
-
-services:
-  app:
-    image: 'docker-spring-boot-postgres:latest'
-    build:
-      context: .
-    ports:
-      - "8080:8080"
-    container_name: app
-    depends_on:
-      - db
-    environment:
-      - SPRING_DATASOURCE_URL=jdbc:postgresql://db:5432/postgres
-      - SPRING_DATASOURCE_USERNAME=postgres
-      - SPRING_DATASOURCE_PASSWORD=1234
-      - SPRING_JPA_HIBERNATE_DDL_AUTO=update
-          
-  db:
-    image: 'postgres:13.2'
-    container_name: db
-    environment:
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=1234
