--- dev/demo/docker_build.sh
@@ -1,4 +1,0 @@
-
-docker build -f Dockerfile -t dockerdemo .
-
-
