--- dev/demo/docker_run.sh
@@ -1,1 +1,0 @@
-docker run dockerdemo
