--- dev/demo/src/main/java/com/demo/DemoApplication.java
@@ -1,23 +1,0 @@
-package com.demo;
-
-
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-
-@SpringBootApplication
-public class DemoApplication {
-
-	
-	
-	public static void main(String[] args){
-		
-		
-		SpringApplication.run(DemoApplication.class, args);
-	}
-
-	
-	
-}
