--- dev/demo/src/main/java/com/demo/controller/JWTController.java
@@ -1,18 +1,0 @@
-package com.demo.controller;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class JWTController {
-
-	
-	
-	
-	
-	
-}
