--- dev/demo/src/main/java/com/demo/model/Usuario.java
@@ -1,61 +1,0 @@
-package com.demo.model;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-import org.springframework.boot.context.properties.bind.DefaultValue;
-
-@Entity
-public class Usuario {
-
-	@Id@GeneratedValue(strategy=GenerationType.AUTO)
-	private int id;
-	private String nombre;
-	private String password;
-	private String token;
-	
-	private String role;
-	
-	public int getId() {
-		return id;
-	}
-	
-	public String getToken() {
-		return token;
-	}
-
-	public void setToken(String token) {
-		this.token = token;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-	
-	public String getNombre() {
-		return nombre;
-	}
-	
-	public void setNombre(String nombre) {
-		this.nombre = nombre;
-	}
-	
-	public String getPassword() {
-		return password;
-	}
-	
-	public void setPassword(String password) {
-		this.password = password;
-	}
-	
-	public void setRole(String role) {
-		this.role = role;
-	}
-	
-	public String getRole() {
-		return role;
-	}
-		
-}
