--- dev/demo/src/main/java/com/demo/repository/PersonaRepo.java
@@ -1,11 +1,0 @@
-package com.demo.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.demo.model.Persona;
-
-public interface PersonaRepo extends JpaRepository<Persona,Integer>{
-
-	
-	
-}
