--- dev/demo/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,5 +1,0 @@
-{"properties": [{
-  "name": "app.api.swagger.enable",
-  "type": "java.lang.String",
-  "description": "A description for 'app.api.swagger.enable'"
-}]}
