--- dev/demo/src/main/resources/application.properties
@@ -1,21 +1,0 @@
-
-
-
-spring.jpa.database=POSTGRESQL
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost/postgres
-spring.datasource.username=postgres
-spring.datasource.password=1234
-spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
-
-
-
-spring.web.resources.add-mappings=true
-
-spring.security.user.name=pablo
-spring.security.user.password=1234
-
-
-app.api.swagger.enable=true
